refactor: split role command up
